Add deploy security token (e.g. db_cmsprod#abcd_token_efgh)

+++ lib/service.moon
@@ -56,19 +56,26 @@ install_service = (sub_domain, name)->
   )
 --------------------------------------------------------------------------------
 deploy_site = (sub_domain, settings) ->
+  deploy_to = stringy.split(sub_domain, "#")
   config = require('lapis.config').get!
   db_config = require('lapis.config').get("db_#{config._name}")
-  path = "dump/#{sub_domain}/"
-  os.execute("mkdir -p #{path}")
+  path = "dump/#{sub_domain[1]}/"
   home = from_json(settings.home)
 
-  command = "arangodump --collection layouts --collection partials --collection components --collection spas --collection redirections --collection datatypes --collection aqls --collection helpers --collection apis --collection api_libs --collection api_routes --collection api_scripts --collection api_tests --collection sripts --collection pages --collection trads --collection uploads --collection folder_path --collection folders --collection scripts --include-system-collections true --server.database db_#{sub_domain} --server.username #{db_config.login} --server.password #{db_config.pass} --server.endpoint #{db_config.endpoint} --output-directory #{path} --overwrite true"
-  command ..= " --collection datasets" if home['deploy_datasets']
+  request = 'FOR settings IN settings LIMIT 1 RETURN setting'
+  sub_domain_settings = aql("db_#{sub_domain}", request)[1]
 
-  os.execute(command)
+  if deploy_to[2] == sub_domain_settings.token
+    os.execute("mkdir -p #{path}")
 
-  os.execute("arangorestore --include-system-collections true --server.database #{settings.deploy_secret} --server.username #{db_config.login} --server.password #{db_config.pass} --server.endpoint #{db_config.endpoint}  --input-directory #{path} --overwrite true")
-  os.execute("rm -Rf #{path}")
+    command = "arangodump --collection layouts --collection partials --collection components --collection spas --collection redirections --collection datatypes --collection aqls --collection helpers --collection apis --collection api_libs --collection api_routes --collection api_scripts --collection api_tests --collection sripts --collection pages --collection trads --collection uploads --collection folder_path --collection folders --collection scripts --include-system-collections true --server.database db_#{sub_domain} --server.username #{db_config.login} --server.password #{db_config.pass} --server.endpoint #{db_config.endpoint} --output-directory #{path} --overwrite true"
+    command ..= " --collection datasets" if home['deploy_datasets']
+
+    os.execute(command)
+
+    os.execute("arangorestore --include-system-collections true --server.database #{settings.deploy_secret} --server.username #{db_config.login} --server.password #{db_config.pass} --server.endpoint #{db_config.endpoint}  --input-directory #{path} --overwrite true")
+    os.execute("rm -Rf #{path}")
+
 --------------------------------------------------------------------------------
 install_script = (sub_domain, name) ->
   path = "scripts/#{sub_domain}/#{name}"
